requirements.txt
httpx>=0.23.0
Pillow>=10.0.0
requests>=2.28.0
typer>=0.9.0